+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [0.2.0] - 05-04-2020.
+Breaking change, the response object has been refined to a more
+suitable object that contains the status of the api and the error message.
+## [0.2.0] - 05-04-2020.
 Add travel mode,
 Fixed unhandled error exception.
 ## [0.1.0] - 25-04-2019.